Allow options.links to inherit prototypically
With ES5 `Object.create`, it becomes possible for an object to be used as a dictionary (with the exception of `__proto__` until ES6).  It furthermore becomes possible for these to be scoped dictionaries, inheriting prototypically.  This would be particularly useful as a data structure for the relevant link for a term in a particular scope.

I propose a relaxation on the "ownership" requirement on the links option.  By testing instead of presence and string type, false positives from the `Object.prototype` are avoided for users not currently using `Object.create`.